fix(display): use integer math for splash row count
The all-constant int((float64(ScreenH)-top)/lineH) expression is folded
at compile time and Go refuses to truncate the fractional float constant
to int. Compute rows with integer division instead.

+++ internal/display/render.go
@@ -366,7 +366,7 @@ func renderSplash(ver string, lines []string) *image.RGBA {
 		fontPx  = 10.0
 	)
 	maxW := float64(ScreenW) - 2*leftPad
-	rows := int((float64(ScreenH) - top) / lineH)
+	rows := (ScreenH - top) / int(lineH) // integer math: avoids a const float→int truncation
 	if len(lines) > rows {
 		lines = lines[len(lines)-rows:]
 	}
